[Refactor] service 구조 리팩토링
