+++ app/src/main/java/org/android/go/sopt/remote/ApiFactory.kt
@@ -5,10 +5,8 @@ import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import org.android.go.sopt.BuildConfig.AUTH_BASE_URL
 import org.android.go.sopt.BuildConfig.USERS_LIST_BASE_URL
-import org.android.go.sopt.remote.service.UsersListService
-import org.android.go.sopt.remote.service.LogInService
-import org.android.go.sopt.remote.service.MyProfileService
-import org.android.go.sopt.remote.service.SignUpService
+import org.android.go.sopt.remote.service.LogInPageService
+import org.android.go.sopt.remote.service.MainPageService
 import retrofit2.Retrofit
 
 object ApiFactory {
@@ -32,8 +30,6 @@ object ApiFactory {
 
 
 object ServicePool { // 서비스가 모여있는곳
-    val signUpService = ApiFactory.createAuth<SignUpService>()
-    val logInService = ApiFactory.createAuth<LogInService>()
-    val listUsersService = ApiFactory.createUsers<UsersListService>()
-    val myProfileService = ApiFactory.createAuth<MyProfileService>()
+    val loginPageService = ApiFactory.createAuth<LogInPageService>()
+    val mainPageService = ApiFactory.createUsers<MainPageService>()
 }
