+++ app/src/main/java/org/android/go/sopt/remote/service/MainPageService.kt
@@ -1,12 +1,14 @@
 package org.android.go.sopt.remote.service
 
-import kotlinx.serialization.Serializable
 import org.android.go.sopt.remote.remoteData.model.MyProfileDto
+import org.android.go.sopt.remote.remoteData.model.ResponseListUsersDto
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
 
-interface MyProfileService {
-    @GET("info/{userId}")
-    fun myProfile(@Path("userId") userId: String):Call<MyProfileDto>
+interface MainPageService {
+
+    @GET("users")
+    fun getListUsers(
+    ): Call<ResponseListUsersDto>
 }
